[Warningfix] [VT-Client] [Removed warning due to comparison]

+++ library/xgpl_src/IsoAgLib/comm/Part6_VirtualTerminal_Client/impl/vtclientconnection_c.cpp
@@ -964,9 +964,9 @@ VtClientConnection_c::getVtObjectPoolSoftKeyHeight() const
 uint8_t
 VtClientConnection_c::getVersion() const
 {
-  return ( getVtServerInst().getVtIsoVersion() < getPool().getVersion() )
+  return ( getVtServerInst().getVtIsoVersion() < uint8_t( getPool().getVersion() ) )
          ? getVtServerInst().getVtIsoVersion()
-         : getPool().getVersion();
+         : uint8_t( getPool().getVersion() );
 }
